+++ web_v2/handlers/get_poll.go
@@ -41,6 +41,10 @@ func GetPoll(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if !isAllowed(r, w, result.Data.Relationships.Owner.Data.ID, result.Data.Relationships.ResultProvider.Data.ID) {
+		return
+	}
+
 	ape.Render(w, result)
 }