+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 * `GetBalanceList` handler to handle `v3/balances` request
 * `GetPublicKeyEntry` handler to handle `v3/public_key_entries/{id}` request
 * Receiver filter for create issuance requests
+* Polls and votes ingestion
+* `GetPollList` to handle `v3/polls`
+* `GetPoll` to handle `v3/polls/{id}`
+* `GetVoteList` to handle `v3/polls/{id}/relationships/votes`
+* `GetVote` to handle `v3/polls/{id}/relationships/votes{voter}`
 
 ### Fixed